Move Espresso helpers to test-shared
