+++ collect_app/src/androidTest/java/org/odk/collect/android/support/pages/NotificationDrawer.kt
@@ -10,7 +10,7 @@ import org.hamcrest.CoreMatchers.`is`
 import org.hamcrest.CoreMatchers.not
 import org.hamcrest.CoreMatchers.nullValue
 import org.hamcrest.MatcherAssert.assertThat
-import org.odk.collect.android.support.WaitFor.waitFor
+import org.odk.collect.testshared.WaitFor.waitFor
 
 class NotificationDrawer {
     private var isOpen = false
